+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Paste nsec login [PR #59](https://github.com/marmot-protocol/sloth/pull/59)
 - Reactions [PR #60](https://github.com/marmot-protocol/sloth/pull/60)
 - Add multi-account support [PR #78](https://github.com/marmot-protocol/sloth/pull/78)
+- Emoji picker for reactions [PR #81](https://github.com/marmot-protocol/sloth/pull/81)
 
 
 ### Changed